inserting direct messaging in app
