+++ index.html
@@ -1,0 +1,23 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+    <link rel="stylesheet" href="/css/style.css">
+</head>
+<body>
+    <div class="flex-column justify-space-around">
+        <header class="display-flex justify-space-between align-center p-2">
+            <h5><a href="/message">Direct message me</a></h5>
+        </header>
+    </div>
+    <form class="form message-form">
+        <div class="form-group">
+          <label for="message-input">Message</label>
+          <input class="form-input" type="text" id="message-input" />
+          <button class="btn submit-btn" type="submit">Send</button>
+        </div>
+  </form>
+</body>
+</html>
