+++ models/Request.js
@@ -23,10 +23,6 @@ Request.init(
       allowNull: false,
       defaultValue: DataTypes.NOW,
     },
-    needed_funding: {
-      type: DataTypes.FLOAT,
-      allowNull: false,
-    },
     owner_id: {
       type: DataTypes.INTEGER,
       references: {
