+++ package.json
@@ -19,6 +19,7 @@
     "express-session": "^1.17.1",
     "inquirer": "^6.3.1",
     "mysql2": "^2.2.5",
-    "sequelize": "^6.6.2"
+    "sequelize": "^6.6.2",
+    "socket.io": "^4.0.1"
   }
 }
