--- public/js/login.js
@@ -1,50 +1,0 @@
-const loginForm = async (event) => {
-    event.preventDefault();
-  
-    const email = document.querySelector('#email-login').value.trim();
-    const password = document.querySelector('#password-login').value.trim();
-  
-    if (email && password) {
-      const response = await fetch('/api/owners/login', {
-        method: 'POST',
-        body: JSON.stringify({ email, password }),
-        headers: { 'Content-Type': 'application/json' },
-      });
-  
-      if (response.ok) {
-        document.location.replace('/schedule');
-      } else {
-        console.log('where are u')
-        alert(response.statusText);
-      }
-    }
-};
-  
-const signupForm = async (event) => {
-    event.preventDefault();
-  
-    const name = document.querySelector('#name-signup').value.trim();
-    const email = document.querySelector('#email-signup').value.trim();
-    const password = document.querySelector('#password-signup').value.trim();
-  
-    if (name && email && password) {
-      const response = await fetch('/api/owners', {
-        method: 'POST',
-        body: JSON.stringify({ name, email, password }),
-        headers: { 'Content-Type': 'application/json' },
-      });
-  
-      if (response.ok) {
-        document.location.replace('/schedule');
-      } else {
-        alert(response.statusText);
-      }
-    }
-};
-  
-document.querySelector('.login-form')
-document.addEventListener('click', loginForm);
-  
-document.querySelector('.signup-form')
-document.addEventListener('click', signupForm);
-
