+++ views/homepage.handlebars
@@ -40,4 +40,4 @@
 </div>
 
 
-<script src="./js/login.js"></script>
+<script src="./js/homepage.js"></script>
