+++ views/layouts/partials/message.handlebars
@@ -1,0 +1,8 @@
+<form class="form message-form" id="send-container">
+      <div class="form-group">
+        <input class="form-input" type="text" id="message-input" />
+        <button class="btn btn-primary" type="submit">Send</button>
+      </div>
+</form>
+
+<script src="/js/message.js"></script>
