+++ views/schedule.handlebars
@@ -1,3 +1,5 @@
-<div class="row">
-    <p>Hello</p>
+<div class="flex-column justify-space-around">
+        <header class="display-flex justify-space-between align-center p-2">
+            <h5><a href="/message">Direct message me</a></h5>
+        </header>
 </div>
